Fixing lectura de asistencia

+++ src/main/java/com/sebasgoy/constantes/Modalidades.java
@@ -4,8 +4,8 @@ public class Modalidades {
 
 	
 	public static String  LIBRE = "LIBRE";
-	public static int  ID_LIBRE = 2;
+	public static Long  ID_LIBRE = 2L;
 
 	public static String  MODULO = "MODULO";
-	public static int  ID_MODULO = 1;
+	public static Long ID_MODULO = 1L;
  }
